move sorting into sorting.js, move sorting tests into sorting.spec.js, add method comments
